requirements.txt
ultralytics
opencv-python
torch
torchvision
torch
torchvision
faiss-cpu
numpy
fastapi
uvicorn
streamlit
requests
python-multipart
pillow